while loop on recv
I failed my exam a few times on test 4 or 5 (the one when they send a file of 100k+ characters). I failed because the recv(fd, str, sizeof(str), 0) only reads about 46k characters, that we lose because the '\n' was not send yet.
I corrected this by doing a loop on recv with a fixed buffer (we don't care about the buffer size). Failed on test 9 (when they do 1k+ clients), pushed it again with no changes and it worked.
I validated this morning.